rename directories and clean-up before refactoring
